refactor(eta-service): reorganize directory structure to match G2 standard
