+++ services/eta-service/app/database/connection.py
@@ -6,10 +6,10 @@ from typing import Generator
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session, sessionmaker
 
-from config import settings
-from db.models import Base
+from app.config import settings
+from app.database.models import Base
 
-engine = create_engine(settings.database_url, pool_pre_ping=True, future=True)
+engine = create_engine(settings.eta_database_url, pool_pre_ping=True, future=True)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
