+++ services/eta-service/app/prediction/inference_router.py
@@ -10,8 +10,8 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Optional
 
-from models._repo_root import repo_root
-from models.eta import EtaResult, compute_eta, compute_eta_expressway
+from app.prediction._repo_root import repo_root
+from app.prediction.eta import EtaResult, compute_eta, compute_eta_expressway
 
 logger = logging.getLogger(__name__)
 
@@ -110,7 +110,7 @@ def _predict_xgboost(
     dt: Optional[datetime.datetime],
     segment_mode: str,
 ) -> InferenceOutcome:
-    from models.ml_eta_xgb import predict_eta_xgb
+    from app.prediction.ml_eta_xgb import predict_eta_xgb
 
     registered_name = _segment_model_name(segment_mode)
     result = predict_eta_xgb(
@@ -121,7 +121,7 @@ def _predict_xgboost(
         registered_name=registered_name,
     )
     used = "xgboost" if not result.clamped else "physics"
-    from models.ml_eta_xgb import get_last_model_metadata
+    from app.prediction.ml_eta_xgb import get_last_model_metadata
 
     meta = get_last_model_metadata()
     return InferenceOutcome(
@@ -147,7 +147,7 @@ def _predict_sarima(
     if not route_id or stop_id is None:
         return None
     try:
-        from models.sarima_eta import forecast_eta_sarima
+        from app.prediction.sarima_eta import forecast_eta_sarima
 
         seconds = forecast_eta_sarima(route_id, int(stop_id))
         if seconds is None:
