--- services/eta-service/db/__init__.py
@@ -1,4 +1,0 @@
-from db.models import Base, EtaRecord
-from db.session import get_session, init_db
-
-__all__ = ["Base", "EtaRecord", "get_session", "init_db"]
