--- services/eta-service/routers/__init__.py
@@ -1,1 +1,0 @@
-# Package marker for routers
