+++ services/eta-service/tests/unit/test_consumer.py
@@ -10,8 +10,8 @@ SERVICE_ROOT = Path(__file__).resolve().parents[2]
 if str(SERVICE_ROOT) not in sys.path:
     sys.path.insert(0, str(SERVICE_ROOT))
 
-from consumer import ETA_LIVE_CHANNEL, EtaFeatureConsumer
-from models.inference_router import InferenceOutcome
+from app.consumers.eta_consumer import ETA_LIVE_CHANNEL, EtaFeatureConsumer
+from app.prediction.inference_router import InferenceOutcome
 
 
 class FakeRedis:
